Extract base class for common code generator tests.
